backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.4
langchain==0.2.5
langchain-openai==0.1.8
openai==1.35.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
PyMuPDF==1.24.5
sentence-transformers==3.0.1
tiktoken==0.7.0
tqdm==4.66.4
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
